feat: SeriesFormModal을 새로운 Modal 구조로 수정
