+++ client/components/shared/Modals/SeriesFormModal/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SeriesFormModal";
