+++ client/components/shared/Modals/SeriesFormModal/styles.module.scss
@@ -1,0 +1,18 @@
+@import "utils/utils.scss";
+
+.Form {
+  display: flex;
+  flex-direction: column;
+  @include mobile {
+    display: flex;
+    flex-direction: column;
+  }
+  input {
+    @include StyledInput;
+    margin-bottom: 10px;
+  }
+  button {
+    @include StyledButton;
+    width: 100%;
+  }
+}
