+++ client/tests/posts.e2e.test.ts
@@ -149,7 +149,10 @@ test.describe("모달 - ", () => {
 
     await posts.mockGetUserAPI();
 
-    await posts.page.getByRole("button", { name: "로그인" }).click();
+    await posts.page
+      .locator("form")
+      .getByRole("button", { name: "로그인" })
+      .click();
 
     await expect(
       posts.page.getByText(`${USER_MOCK_DATA.nickname}님`)
